Fix double scrollbars in multiepg views
When showing multiepg views, the contents are often presented
in a view that has scrollbars for the EPG table, but this view
is subsequently embedded in another view with scrollbars. This
results in double scrollbars and a user interface that is very
difficult to navigate.

There is actually no need to specify an explicit height for the
table body. Removing this height gets rid of the inner scrollbar
and as a result the view only ends up with the outer scrollbar.


Signed-off-by: Peter Urbanec <git.user@urbanec.net>