[WebTV] Enhance interaction with auth It's now also possible to view recordings in WebTV with auth for streaming enabled.
Also modified URL generation a bit to increase compatibility with PLi streaming:
PLi streaming doesn't support auth in WebTV and there is nothing we can do about it.
WebTV on distros using PLi streaming will be restricted to auth for streaming disabled or WebTV will only work for recordings (And on some machines not even that when used with transcoding).