[Streaming Auth] Use basic auth Uses a volatile "user account" "-sid" with session id as password. Works with file streaming and live streaming. It works for transcoding using VLC but not for VXG.
It does not work for PLi streaming:
Does not auth through OWIF and the user "-sid" doesn't really exist.
"-sid" is intentionally even non-POSIX-compliant, so it CAN NOT exist and thus not collide.

auth is not appended if streaming auth is disabled, but even when it accidently is there is no problem:
Streaming doesn't care about credentials unless they are required.