grab.py: Allow multiple grab instances to run at the same time
Only one instance of "grabScreenshot" exists in the system. When two grabs are
run in parallel, the request member would get replaced with a new one, and when
data or finished events arrive from the previous request, the system gets confused
and starts throwing exceptions.

To solve the issue, create a handler object for each request, keep it alive for
as long as the request exists and create an eConsoleAppContainer (another class
that doesn't handle multiple invokations well) for it to use exclusively. This
makes the server capable of handling multiple requests without crashing.

Solves: #559 "Crash when creating screenshots in quick consecution"