grab.py: Implement "notifyFinish" to handle disconnected clients
Holding down the F5 key on a grab image will make the browser close the connection
to the server. The request then emits a notifyFinish() which must then abort the
request and the request.finish() must not be called after that.

This solves the following error message from twisted:
Request.finish called on a request after its connection was lost; use
Request.notifyFinish to keep track of this.

Solves: #559