Scale Correctly the SNR/AGC values
According to the documentation the drivers return a value from 0 to 0xffff (100%).

https://linuxtv.org/downloads/v4l-dvb-apis/uapi/dvb/frontend-stat-properties.html

In some places we are using 65536 (0x10000) instead of 65535 (0xffff).
That was causing the maximum value to displayed as 99% instead of 100%.

Replace the 65536 with 65535 in frontend CNR/SNR calculatiosn to fix issue.

Following changes in: https://github.com/OpenPLi/enigma2/pull/656