again changed height size from 400px to 450px
Sorry for that, but I can not test it in the normal way ! If it is wrong again, please test and correct somebody. I also think ui-button style is very narrow (there is not enough text) and would like to adjust the style of these buttons - to increase their width. Thanks.