httpserver: always allow /web/stream?StreamService= url from localhost.
When access is from localhost and requesting stream?StreamService
always allow it (not authentication required).

This is required by external applications (e.g. streamproxy)
that require enigma2 to setup a service (tuner/demux). The URL
in itself does not allow streaming so it is safe to always allow
it to localhost.