Separation of concerns - modern & legacy assets
This change isolates js code between the modern and legacy interfaces.
As part of this, jQuery 2.2.4 is used, which is equivalent to 1.12.4,
but removes support for IE 6/7/8.

This change does not upgrade to latest version of jQuery (3.x) as
this is likely to introduce breaking changes.
Attention instead should (in the case of modern interface) be
focussed on removing jQuery dependence altogether.