Simplify process of finding matching timers in Multi EPG.
It seems that events and timers are always ordered by begin time (if not: they can easily be sorted).
In that case we can always pick the first timer from the timers list and check if it belongs to the currently processed event.
If the processed event begins later than this timer, remove it from the list.